// src/server/shared/enums/EFilaName.ts
export enum EFilaNames {
    filaVisita = 'fila-visita',
    filaIpBlock = 'fila-ip-block',
    filaClick = 'fila-click',
    filaVisitaMetrica = 'fila-visita-metrica',
    postbackLinkEnvios = 'postback-link-envios',
    postbackLinkEnviosIntegracao = 'postback-link-envios-integracao',
    checkout = 'checkout',
    checkoutIntegracao = 'checkout-integracao'
}